Extract/update OmekaGlobals
Reconstruct a working version of the hacked-up OmekaGlobals script,
including the minimal necessary bits of sphpdox and altering it to use
PHP native reflection.

Also remove some system-specific hardcoded paths and barriers to current
usage like that.

This does not include the "packages" code: generated files will
reference their packages but the indexes will not be updated
automatically.